Refactored oauth controllers file tree
