+++ ratify-be/controllers/oauth/authorize.go
@@ -1,0 +1,75 @@
+package oauth
+
+import (
+	"fmt"
+	"net/http"
+	"strings"
+
+	"github.com/gin-gonic/gin"
+	"github.com/google/go-querystring/query"
+
+	"github.com/daystram/ratify/ratify-be/constants"
+	"github.com/daystram/ratify/ratify-be/datatransfers"
+	"github.com/daystram/ratify/ratify-be/handlers"
+	"github.com/daystram/ratify/ratify-be/models"
+)
+
+// @Summary Request authorization
+// @Tags oauth
+// @Param user body datatransfers.AuthorizationRequest true "Authorization request info"
+// @Success 200 "OK"
+// @Router /oauth/authorize [POST]
+func POSTAuthorize(c *gin.Context) {
+	var err error
+	// fetch request info
+	var authRequest datatransfers.AuthorizationRequest
+	if err = c.ShouldBindJSON(&authRequest); err != nil {
+		c.JSON(http.StatusBadRequest, datatransfers.APIResponse{Error: err.Error()})
+		return
+	}
+	// retrieve application
+	var application models.Application
+	if application, err = handlers.Handler.RetrieveApplication(authRequest.ClientID); err != nil {
+		c.JSON(http.StatusNotFound, datatransfers.APIResponse{Error: "application not found"})
+		return
+	}
+	flow := authRequest.Flow()
+	switch flow {
+	case constants.FlowAuthorizationCode, constants.FlowAuthorizationCodeWithPKCE:
+		// verify user login
+		var user models.User
+		if user, err = handlers.Handler.AuthenticateUser(authRequest.UserLogin); err != nil {
+			c.JSON(http.StatusUnauthorized, datatransfers.APIResponse{Code: "incorrect_credentials", Error: "incorrect username or password"})
+			return
+		}
+		// verify request credentials
+		// TODO: support comma-separated callback URLs
+		if authRequest.RedirectURI != "" && authRequest.RedirectURI != application.CallbackURL {
+			c.JSON(http.StatusUnauthorized, datatransfers.APIResponse{Error: "not allowed callback_uri"})
+			return
+		}
+		// generate authorization code
+		var authorizationCode string
+		if authorizationCode, err = handlers.Handler.GenerateAuthorizationCode(authRequest, user.Subject); err != nil {
+			c.JSON(http.StatusUnauthorized, datatransfers.APIResponse{Error: "failed generating authorization_code"})
+			return
+		}
+		// note code challenge
+		if flow == constants.FlowAuthorizationCodeWithPKCE {
+			if err = handlers.Handler.StoreCodeChallenge(authorizationCode, authRequest.PKCEAuthFields); err != nil {
+				c.JSON(http.StatusUnauthorized, datatransfers.APIResponse{Error: "failed storing code_challenge"})
+				return
+			}
+		}
+		param, _ := query.Values(datatransfers.AuthorizationResponse{
+			AuthorizationCode: authorizationCode,
+			State:             authRequest.State,
+		})
+		c.JSON(http.StatusOK, gin.H{
+			"data": fmt.Sprintf("%s?%s", strings.TrimSuffix(application.CallbackURL, "/"), param.Encode()),
+		})
+	default:
+		c.JSON(http.StatusBadRequest, datatransfers.APIResponse{Error: "unsupported authorization flow"})
+		return
+	}
+}
